requirements.txt

Flask==2.3.3
Flask-Cors==4.0.0
PyJWT==2.8.0
openai==0.28
anthropic==0.8.1
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
httpx==0.27.2
azure-core==1.24.0
azure-identity==1.10.0
requests==2.31.0
selenium
webdriver-manager
browser_cookie3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
presidio-analyzer
presidio-anonymizer
spacy